Fix Add to cart API

+++ swigatto/src/main/java/com/example/swigatto/model/Cart.java
@@ -23,7 +23,7 @@ public class Cart {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
 
-    int cartTotal;
+    double cartTotal;
 
     @OneToOne
     @JoinColumn
@@ -31,4 +31,6 @@ public class Cart {
 
     @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL)
     List<FoodItem> foodItems = new ArrayList<>();
+
+
 }
